requirements.txt
streamlit==1.38.0
chromadb==0.5.5
pandas==2.2.2
pyyaml==6.0.2
openai==1.40.6
tqdm==4.66.5
torch==2.2.2
sentence-transformers==2.6.1
transformers==4.41.2
